requirements.txt
fastapi
mypy
black
pydantic
uvicorn
langchain
openai
python-dotenv
